requirements.txt
gradio
google-generativeai
langchain-community
langchain
langchain-google-genai
faiss-cpu
pymupdf
sentence-transformers
rank_bm25
torch